requirements.txt
colorama==0.4.6
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.4
solana==0.36.5
solders==0.25.0
Django==5.1.6
gunicorn==23.0.0
aiogram==3.18.0
python-dotenv==1.0.1
pycryptodome==3.21.0
watchdog==6.0.0
aiohttp==3.11.12
pydantic==2.10.6
